+++ pattern_tree.h
@@ -8,12 +8,12 @@ pattern_node create_new_pattern_node(data_type);
 
 void create_and_insert_new_child(pattern_node, data);
 
-pattern_node insert_itemset_helper(pattern_node, data, int, int);
-patterntree insert_itemset(patterntree, data, int, int);
+pattern_node insert_itemset_helper(pattern_node, data, int, float);
+patterntree insert_itemset(patterntree, data, int, float);
 
-void update_tilted_tw_table(pattern_node, int, int);
-tilted_tw_table insert_batch(tilted_tw_table, int, int, int);
-tilted_tw_table create_new_tilted_tw_table(int, int, int);
+void update_tilted_tw_table(pattern_node, int, float);
+tilted_tw_table insert_batch(tilted_tw_table, int, int, float);
+tilted_tw_table create_new_tilted_tw_table(int, int, float);
 
 fpnode dfs(pattern_node);
 fptree get_fptree(patterntree);
